option: merge option.c to libtrx
