+++ src/tr1/meson.build
@@ -108,7 +108,6 @@ sources = [
   'game/inventory.c',
   'game/objects/common.c',
   'game/objects/setup.c',
-  'game/option/option.c',
   'game/option/option_compass.c',
   'game/savegame/savegame_bson.c',
   'game/savegame/savegame_legacy.c',
