--- src/tr2/game/option/option.h
@@ -1,3 +1,0 @@
-#pragma once
-
-#include <libtrx/game/option/common.h>
