+++ src/tr2/game/option/option_compass.c
@@ -1,4 +1,4 @@
-#include "game/option/option.h"
+#include "game/option/option_compass.h"
 #include "game/savegame.h"
 
 #include <libtrx/game/game.h>
