+++ src/tr2/game/option/option_compass.h
@@ -1,0 +1,3 @@
+#pragma once
+
+#include <libtrx/game/option/compass.h>
