+++ src/tr2/meson.build
@@ -103,7 +103,6 @@ sources = [
   'game/objects/general/harpoon_bolt.c',
   'game/objects/general/zipline.c',
   'game/objects/setup.c',
-  'game/option/option.c',
   'game/option/option_compass.c',
   'game/savegame/savegame_bson.c',
   'game/savegame/savegame_legacy.c',
